% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateChunks.R
\name{calculateChunks}
\alias{calculateChunks}
\title{Calculates number of chunks, returns chunk head/tail indices in a list.}
\usage{
calculateChunks(len, threshold)
}
\description{
This function calculates the number of chunks to make, and returns
head/tail index pairs c(head,tail) in a list list(c(head1,tail1) ... c(headn,tailn))
the maximum length of a chunk is given by the threshold parameter, which
should be set according to the size of the data in each row, and the
intensity of the calculation.
}
\examples{
len <- nrow(mtcars)
chunks <- calculateChunks(len,10)\%>\%
	chunkIndices()

mtcars[chunkIndices[[1]],]

}
\keyword{bigdata}
