% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkIndices.R
\name{chunkIndices}
\alias{chunkIndices}
\title{Returns the indices of a list of chunk pairs (from calculateChunks)}
\usage{
chunkIndices(chunks)
}
\description{
This is a helper function that returns a list of all of the indices within each chunk.
}
\examples{
len<- nrow(mtcars)
chunks <- calculateChunks(len,10)\%>\%
	chunkIndices()

mtcars[chunkIndices[[1]],]

}
\keyword{bigdata}
