% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_lodJson.R
\name{df_to_lodJson}
\alias{df_to_lodJson}
\title{Dump a data frame to a vector of json-formatted dictionaries}
\usage{
df_to_lodJson(df)
}
\description{
This function pops entries from a redis list, either until the key returns NULL 
or until it has read n keys. Similar to readLines
}
\examples{
json <- df_to_lodJson(mtcars)

}
\keyword{path}
