% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfToVector.R
\name{dfToVector}
\alias{dfToVector}
\title{Utility function that converts a data frame into a character vector.}
\usage{
dfToVector(df, sanitize = FALSE, ...)
}
\description{
This function takes a data frame, and returns csv-formatted lines in a vector.
}
\examples{



}
\keyword{path}
