% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lodJson.R
\name{read_lodJson}
\alias{read_lodJson}
\title{Read a data frame from a vector of json-formatted dictionaries}
\usage{
read_lodJson(lod)
}
\description{
This function pops entries from a redis list, either until the key returns NULL 
or until it has read n keys. Similar to readLines
}
\examples{
lod <- df_to_lodJson(mtcars)
newmtcars <- read_lodJson(lod)

}
\keyword{path}
