% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCsvVector.R
\name{readCsvVector}
\alias{readCsvVector}
\title{Reads a dataset from a character vector}
\usage{
readCsvVector(vec, nlsub = " ", ...)
}
\description{
This function reads a dataset from a character vector, where each
value is a row in csv-format.
}
\details{
NOTE! This function uses the \\n character to unserialize the data. 
Any newline characters in the data are substituted by the nlsub
character.
}
\examples{
rep <- capture.output(write.csv(mtcars)) 
dat <- readCsvVector(rep) 

}
\keyword{path}
