% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redisChunkApply.R
\name{redisChunkApply}
\alias{redisChunkApply}
\title{Applies FUN to data stored in a redis database}
\usage{
redisChunkApply(redis, key, FUN, chunksize = 100, verbose = FALSE,
  sanitize = FALSE, ...)
}
\description{
This function applies FUN to data stored in a redis database, chunking the operation if
number of rows exceeds the chunksize. 
NOTE: this function uses (wipes) the '{key}_tmp' key in the database.
}
\examples{
r <- redis()
processRedisData(r,'data',as.sentence.df,'body') 

}
\keyword{bigdata}
