% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redisGet.R
\name{redisGet}
\alias{redisGet}
\title{Utility function that reads lines from a redis list.}
\usage{
redisGet(redis, key, n = NULL)
}
\description{
This function pops entries from a redis list, either until the key returns NULL 
or until it has read n keys. Similar to readLines
}
\examples{



}
\keyword{path}
